+++ openwisp_monitoring/monitoring/management/commands/__init__.py
@@ -1,0 +1,10 @@
+from django.core.management.base import BaseCommand
+
+from ...tasks import migrate_timeseries_database
+
+
+class BaseMigrateTimeseriesCommand(BaseCommand):
+    help = 'Migrate timeseries database asynchronously'
+
+    def handle(self, *args, **options):
+        migrate_timeseries_database.delay()