+++ CHANGES.rst
@@ -18,6 +18,20 @@ Backward incompatible changes
 - If you have made changes to the default *Monitoring Template*, then
   create a backup of your template before running migrations. Running
   migrations will make changes to the default *Monitoring Template*.
+- The time-series database schema for storing
+  `interface traffic <https://github.com/openwisp/openwisp-monitoring#traffic>`_
+  and `associated WiFi clients <https://github.com/openwisp/openwisp-monitoring#wifi-clients>`_
+  has been updated. The data for *interface traffic* and *associated WiFi clients*
+  is stored in ``traffic`` and ``wifi_clients`` measurements respectively.
+  The Django migrations will perform the necessary operations in the time-series
+  database aysnchronously. It is recommended that you backup the time-series
+  database before running the migrations.
+
+  You can use the `migrate_timeseries <https://github.com/openwisp/openwisp-monitoring#run-checks>`_
+  management command to trigger the migration of the time-series database.
+- The `interface traffic <https://github.com/openwisp/openwisp-monitoring#traffic>`_
+  and `associated WiFi clients <https://github.com/openwisp/openwisp-monitoring#wifi-clients>`_
+  metrics store additional tags, i.e. ``organization_id``, ``location_id`` and ``floorplan_id``.
 
 Version 0.1.0 [2021-01-31]
 --------------------------